Moved functions from veneer.c/.h and object.c/.h
