+++ src/classes/CMakeLists.txt
@@ -5,6 +5,7 @@ target_sources(morpho
         cmplx.c    cmplx.h
         dict.c     dict.h
         file.c     file.h
+        function.c function.h
         list.c     list.h
         range.c    range.h
         strng.c    strng.h
