+++ src/datastructures/object.h
@@ -147,55 +147,6 @@ objectclass *object_newclass(value name);
 objectclass *morpho_lookupclass(value obj);
 
 /* ---------------------------
- * Functions
- * --------------------------- */
-
-extern objecttype objectfunctiontype;
-#define OBJECT_FUNCTION objectfunctiontype
-
-typedef struct {
-    value symbol; /** Symbol associated with the variable */
-    indx def; /** Default value as constant */
-    indx reg; /** Associated register */
-} optionalparam;
-
-DECLARE_VARRAY(optionalparam, optionalparam)
-
-/** A function object */
-typedef struct sobjectfunction {
-    object obj;
-    int nargs;
-    int varg; // The parameter number of a variadic parameter.
-    value name;
-    indx entry;
-    struct sobjectfunction *parent;
-    int nupvalues;
-    int nregs;
-    objectclass *klass; 
-    varray_value konst;
-    varray_varray_upvalue prototype;
-    varray_optionalparam opt;
-} objectfunction;
-
-/** Gets an objectfunction from a value */
-#define MORPHO_GETFUNCTION(val)   ((objectfunction *) MORPHO_GETOBJECT(val))
-
-/** Tests whether an object is a function */
-#define MORPHO_ISFUNCTION(val) object_istype(val, OBJECT_FUNCTION)
-
-void object_functioninit(objectfunction *func);
-void object_functionclear(objectfunction *func);
-bool object_functionaddprototype(objectfunction *func, varray_upvalue *v, indx *ix);
-objectfunction *object_getfunctionparent(objectfunction *func);
-value object_getfunctionname(objectfunction *func);
-varray_value *object_functiongetconstanttable(objectfunction *func);
-objectfunction *object_newfunction(indx entry, value name, objectfunction *parent, unsigned int nargs);
-bool object_functionhasvargs(objectfunction *func);
-void object_functionsetvarg(objectfunction *func, unsigned int varg);
-
-void objectfunction_printfn(object *obj);
-
-/* ---------------------------
  * Upvalue objects
  * --------------------------- */
 
@@ -212,10 +163,10 @@ typedef struct sobjectupvalue {
 void object_upvalueinit(objectupvalue *c);
 objectupvalue *object_newupvalue(value *reg);
 
-/** Gets an objectfunction from a value */
+/** Gets an upvalue from a value */
 #define MORPHO_GETUPVALUE(val)   ((objectupvalue *) MORPHO_GETOBJECT(val))
 
-/** Tests whether an object is a function */
+/** Tests whether an object is an upvalue */
 #define MORPHO_ISUPVALUE(val) object_istype(val, object_upvaluetype)
 
 /* ---------------------------
